// web/src/api/model/resourceResponse.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Prospero API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceResponse } from "./sourceResponse";

export interface ResourceResponse {
    /** @nullable */
    admin_notes?: string | null;
    /** @nullable */
    anchor_sentence_end_number?: number | null;
    anchor_sentence_start_number: number;
    created_at: string;
    /** @nullable */
    editor_note?: string | null;
    id: string;
    is_featured: boolean;
    /** @nullable */
    quoted_text?: string | null;
    resource_type: string;
    sentence_kind: string;
    source?: null | SourceResponse;
    /** @nullable */
    source_location_freeform?: string | null;
    /** @nullable */
    source_page_end?: number | null;
    /** @nullable */
    source_page_start?: number | null;
    /** @nullable */
    verbatim_kind?: string | null;
}